♻️ | refactor: functions -> utils
