+++ src/interactions/buttons/music/play.js
@@ -1,5 +1,5 @@
 import { EmbedBuilder } from "discord.js";
-import createPlayer from "../../../functions/player.js";
+import createPlayer from "../../../utils/player.js";
 
 export default {
   customId: "play",
